constructor

+++ include/HuffmanTree.h
@@ -30,7 +30,7 @@ class HuffmanTree {
         HuffmanTreeNode *getRoot();
         std::array<uint64_t, CHAR_COUNT> getFrequencies(),
             codes(std::array<uint8_t, CHAR_COUNT> &lengths);
-        int depth();
+        size_t depth();
 };
 
 #endif
